requirements.txt
psycopg2-binary
aiogram~=3.10.0
python-dotenv~=1.0.1
